MacOS Utility/requirements.txt
pycryptodome
firebase-admin
certifi
Pillow
google-cloud-firestore